Make home screen paginated.
This is a new pattern for paginated content. In screen reader mode the left/right swipe gesture are used for the basic navigation functionality. So a user typically does a 3 finger swipe to go to the next page. Having the pagebar act as a (virtual) slider that controls the page allows the screen reader to move pages when it detects a 3 finger swipe.