Bug 834117 - Failure when we move an icon quickly over panels r=crdlc a=tef+
We keep the list of icons that we animated so that we can remove the animation
exactly for the same list.

Also refactored dragdrop's `overIconGrid` function. We probably can do more but
I didn't want to do more in a TEF+ bug. I tried to fail fast with boolean values
when before we were sometimes doing expensive checks before checking these bool.

Also fixed a bug occurring when the user tried to move an icon below the dock.

Last but not least, we now don't try to animate when there is only one icon
(that is the icon being dragged) in a page.